+++ packages/@react-spectrum/menu/package.json
@@ -53,6 +53,7 @@
     "@react-spectrum/text": "^3.4.2",
     "@react-spectrum/utils": "^3.9.2",
     "@react-stately/collections": "^3.8.0",
+    "@react-stately/flags": "3.0.0-alpha.1",
     "@react-stately/menu": "^3.5.2",
     "@react-stately/overlays": "^3.5.2",
     "@react-stately/tree": "^3.6.1",
