+++ packages/@react-spectrum/menu/src/MenuDialogTrigger.tsx
@@ -17,6 +17,7 @@ import {ItemProps} from '@react-types/shared';
 import {MenuDialogContext, useMenuStateContext} from './context';
 import {Modal, Popover} from '@react-spectrum/overlays';
 import React, {Key, ReactElement, useRef} from 'react';
+import {unavailableMenuItems} from '@react-stately/flags';
 import {useOverlayTriggerState} from '@react-stately/overlays';
 
 export interface SpectrumMenuDialogTriggerProps<T> extends ItemProps<T> {
@@ -65,6 +66,9 @@ function MenuDialogTrigger<T>(props: SpectrumMenuDialogTriggerProps<T>): ReactEl
       }
     }
   };
+  if (!unavailableMenuItems()) {
+    return null;
+  }
   return (
     <>
       <MenuDialogContext.Provider value={{isUnavailable, triggerRef}}>{trigger}</MenuDialogContext.Provider>
@@ -103,15 +107,19 @@ MenuDialogTrigger.getCollectionNode = function* getCollectionNode<T>(props: Item
   let [trigger] = React.Children.toArray(props.children) as ReactElement[];
   let [, content] = props.children as [ReactElement, ReactElement];
 
-  yield {
-    element: React.cloneElement(trigger, {...trigger.props, hasChildItems: true}),
-    wrapper: (element) => (
-      <MenuDialogTrigger key={element.key} targetKey={element.key} {...props}>
-        {element}
-        {content}
-      </MenuDialogTrigger>
-    )
-  };
+  if (unavailableMenuItems()) {
+    yield {
+      element: React.cloneElement(trigger, {...trigger.props, hasChildItems: true}),
+      wrapper: (element) => (
+        <MenuDialogTrigger key={element.key} targetKey={element.key} {...props}>
+          {element}
+          {content}
+        </MenuDialogTrigger>
+      )
+    };
+  } else {
+    yield null;
+  }
 };
 
 let _Item = MenuDialogTrigger as <T>(props: ItemProps<T> & {isUnavailable?: boolean}) => JSX.Element;
